+++ .preview/src/components/knob_control/button_control.jsx
@@ -1,0 +1,13 @@
+import { useKnob } from '../../modules/knobs/hooks';
+
+const ButtonControl = ({ id }) => {
+  const { label, handler } = useKnob(id);
+
+  return (
+    <span className="ui-button" onClick={handler}>
+      {label}
+    </span>
+  );
+};
+
+export default ButtonControl;
