+++ package.json
@@ -74,6 +74,7 @@
     "stylelint": "^13.8.0",
     "stylelint-config-nebenan": "^1.6.2",
     "typescript": "^4.1.2",
+    "updeep": "^1.2.0",
     "val-loader": "^2.1.2",
     "webpack": "^5.36.2",
     "webpack-dev-middleware": "^4.1.0",
